chore: Exception 객체 폴더 구조 수정
