--- src/core/errors/conflictException.ts
@@ -1,7 +1,0 @@
-import { HttpException } from './httpException';
-
-export class ConflictException extends HttpException {
-  constructor(message: string = 'Conflict') {
-    super(409, message);
-  }
-}
