--- src/core/errors/forbiddenException.ts
@@ -1,7 +1,0 @@
-import { HttpException } from './httpException';
-
-export class ForbiddenException extends HttpException {
-  constructor(message: string = '') {
-    super(403, message);
-  }
-}
