--- src/core/errors/index.ts
@@ -1,4 +1,0 @@
-export * from './httpException';
-export * from './notFoundException';
-export * from './conflictException';
-export * from './forbiddenException';
