--- src/core/errors/notFoundException.ts
@@ -1,7 +1,0 @@
-import { HttpException } from './httpException';
-
-export class NotFoundException extends HttpException {
-  constructor(message: string = 'Not Found') {
-    super(404, message);
-  }
-}
