+++ src/routes/Comments/service/service.ts
@@ -3,8 +3,8 @@ import { EXCEPTION_MESSAGES } from '@/constants/exceptionMessages';
 import { Comment } from '@/routes/Comments/model/comment';
 import { AUTH_MESSAGES } from '@/constants/authMessages';
 import CommentRepository from '@/routes/Comments/repository/commentRepository';
-import { NotFoundException } from '@/core/errors';
-import { UnauthorizedException } from '@/core/errors/unauthorizedException';
+import { NotFoundException } from '@/core/errors/httpException';
+import { UnauthorizedException } from '@/core/errors/httpException';
 
 export class CommentService {
   constructor(private commentRepository: CommentRepository) {}
